refactor(coedit): rename lib/coediting → lib/coeditor
